mcp/rag_agent/pyproject.toml
[project]
name = "rag_agent"
version = "0.1.0"
description = "RAG agent for LibreChat"
authors = [{name = "Your Name"}]
requires-python = ">=3.10"
dependencies = [
    "langchain",
    "chromadb",
    "langchain-community",
    "langchain-openai",
    "langchain-chroma>=0.1.0",
    "openai",
    "pypdf",
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "mcp[cli]>=1.12.4",
    "httpx>=0.28.1",
    "cryptography>=42.0.0"
]

[tool.setuptools.packages.find]
where = ["."]



